backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pillow==10.4.0
numpy==1.26.4
paddleocr==2.7.3
paddlepaddle==2.6.2
pdf2image==1.17.0
sqlalchemy==2.0.32
alembic==1.13.2
celery==5.4.0
redis==5.0.8
psycopg[binary]==3.2.1
pytest==8.3.3
httpx==0.27.2
filetype==1.2.0
slowapi==0.1.9
prometheus-client==0.20.0
